+++ src/Sulu/Bundle/ContactBundle/DependencyInjection/Configuration.php
@@ -61,6 +61,7 @@ class Configuration implements ConfigurationInterface
                 ->arrayNode('form')
                     ->addDefaultsIfNotSet()
                     ->children()
+                        ->scalarNode('avatar_collection')->defaultValue(1)->end()
                         ->arrayNode('contact')
                             ->addDefaultsIfNotSet()
                             ->children()