+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/accounts/list/main.js
@@ -76,7 +76,9 @@ define([
 
         header: {
             noBack: true,
+
             title: 'contact.accounts.title',
+            underline: false,
 
             toolbar: {
                 buttons: {
@@ -130,7 +132,8 @@ define([
                 view: this.sandbox.sulu.getUserSetting(constants.listViewStorageKey) || 'datagrid/decorators/card-view',
                 viewOptions: {
                     table: {
-                        actionIconColumn: 'name'
+                        actionIconColumn: 'name',
+                        noImgIcon: 'fa-home'
                     },
                     'datagrid/decorators/card-view': {
                         imageFormat: '100x100-inset',