+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/content/column/main.js
@@ -7,7 +7,10 @@
  * with this source code in the file LICENSE.
  */
 
-define(['sulucontent/components/open-ghost-overlay/main'], function(OpenGhost) {
+define([
+    'sulucontent/components/open-ghost-overlay/main',
+    'sulusecurity/services/security-checker'
+], function(OpenGhost, SecurityChecker) {
 
     'use strict';
 
@@ -42,26 +45,13 @@ define(['sulucontent/components/open-ghost-overlay/main'], function(OpenGhost) {
         ACTION_ICON_VIEW = 'fa-eye',
 
         getActionIcon = function(data) {
-            if (!data.hasOwnProperty('permissions')) {
-                return ACTION_ICON_EDIT;
-            }
-
             var actionIcon = '';
-            $.each(data.permissions, function(role, permission) {
-                if ($.inArray(role, this.sandbox.sulu.user.roles) === -1) {
-                    return true;
-                }
 
-                if ($.inArray('edit', permission) !== -1) {
-                    actionIcon = ACTION_ICON_EDIT;
-
-                    return false;
-                } else if ($.inArray('view', permission) !== -1) {
-                    actionIcon = ACTION_ICON_VIEW;
-
-                    return false;
-                }
-            }.bind(this));
+            if (!!data._permissions.edit) {
+                actionIcon = ACTION_ICON_EDIT;
+            } else if (!!data._permissions.view) {
+                actionIcon = ACTION_ICON_VIEW;
+            }
 
             return actionIcon;
         },
@@ -80,6 +70,90 @@ define(['sulucontent/components/open-ghost-overlay/main'], function(OpenGhost) {
                     '<div id="wait-container" style="margin-top: 50px; margin-bottom: 200px; display: none;"></div>'
                 ].join('');
             }
+        },
+
+
+        /**
+         * Enabler for selected items
+         * @param column
+         * @returns {boolean}
+         */
+        hasSelectedEnabler = function(column) {
+            return !!column.hasSelected;
+        },
+
+        /**
+         * Enabler for DELETE
+         * @param column
+         * @returns {boolean}
+         */
+        deleteEnabler = function(column) {
+            return hasSelectedEnabler(column) && SecurityChecker.hasPermission(column.selectedItem, 'delete');
+        },
+
+        /**
+         * Enabler for MOVE
+         * @param column
+         * @return {boolean}
+         */
+        moveEnabler = function(column) {
+            return hasSelectedEnabler(column) && SecurityChecker.hasPermission(column.selectedItem, 'edit');
+        },
+
+        /**
+         * Enabler for COPY
+         * @param column
+         * @returns {boolean}
+         */
+        copyEnabler = function(column) {
+            return hasSelectedEnabler(column) && SecurityChecker.hasPermission(column.selectedItem, 'view');
+        },
+
+        /**
+         * Enaber for ORDER
+         * @param column
+         * @returns {boolean}
+         */
+        orderEnabler = function(column) {
+            var editChildrenPermission = true;
+
+            $.each(column.children, function(id, childItem) {
+                if (!SecurityChecker.hasPermission(childItem, 'edit')) {
+                    editChildrenPermission = false;
+                    return false;
+                }
+            });
+
+            return column.numberItems > 1 && editChildrenPermission && checkParentSecurity(column, 'edit');
+        },
+
+        /**
+         * Enables the add button if the parent allows
+         * @param column
+         * @returns {boolean}
+         */
+        addButtonEnabler = function(column) {
+            return checkParentSecurity(column, 'add');
+        },
+
+        /**
+         * Checks if the selected item in the parent column has the given permission
+         * @param column
+         * @param permission
+         * @returns {boolean}
+         */
+        checkParentSecurity = function(column, permission) {
+            if (!!column.parent) {
+                if (!!column.parent.hasOwnProperty('_permissions')) {
+                    return column.parent._permissions[permission];
+                }
+
+                if (!column.parent.selectedItem) {
+                    return true;
+                }
+            }
+
+            return SecurityChecker.hasPermission(column.parent.selectedItem, permission);
         };
 
     return {
@@ -449,11 +523,12 @@ define(['sulucontent/components/open-ghost-overlay/main'], function(OpenGhost) {
                         resultKey: 'nodes',
                         url: this.getUrl(),
                         actionIcon: getActionIcon.bind(this),
+                        addButton: addButtonEnabler.bind(this),
                         data: [
                             {
                                 id: DELETE_BUTTON_ID,
                                 name: this.sandbox.translate('content.contents.settings.delete'),
-                                enabler: this.hasSelectedEnabler
+                                enabler: deleteEnabler.bind(this)
                             },
                             {
                                 id: 2,
@@ -462,41 +537,23 @@ define(['sulucontent/components/open-ghost-overlay/main'], function(OpenGhost) {
                             {
                                 id: MOVE_BUTTON_ID,
                                 name: this.sandbox.translate('content.contents.settings.move'),
-                                enabler: this.hasSelectedEnabler
+                                enabler: moveEnabler.bind(this)
                             },
                             {
                                 id: COPY_BUTTON_ID,
                                 name: this.sandbox.translate('content.contents.settings.copy'),
-                                enabler: this.hasSelectedEnabler
+                                enabler: copyEnabler.bind(this)
                             },
                             {
                                 id: ORDER_BUTTON_ID,
                                 name: this.sandbox.translate('content.contents.settings.order'),
                                 mode: 'order',
-                                enabler: this.orderEnabler
+                                enabler: orderEnabler.bind(this)
                             }
                         ]
                     }
                 }
             ]);
-        },
-
-        /**
-         * Enabler for MOVE, COPY and DELETE
-         * @param column
-         * @returns {boolean}
-         */
-        hasSelectedEnabler: function(column) {
-            return !!column.hasSelected;
-        },
-
-        /**
-         * Enaber for ORDER
-         * @param column
-         * @returns {boolean}
-         */
-        orderEnabler: function(column) {
-            return column.numberItems > 1;
         },
 
         /**