+++ src/Sulu/Bundle/TranslateBundle/Controller/CodeController.php
@@ -78,9 +78,12 @@ class CodeController extends RestController implements ClassResourceInterface
             $listHelper = $this->get('sulu_core.list_rest_helper');
             $limit = $listHelper->getLimit();
             $offset = $listHelper->getOffset();
-            $sortOrder = $listHelper->getSortOrder();
             $sortColumn = $listHelper->getSortColumn();
-            $sorting = [$sortColumn => $sortOrder];
+            $sorting = ['id' => 'ASC'];
+            if ($sortColumn) {
+                $sortOrder = $listHelper->getSortOrder();
+                $sorting = [$sortColumn => $sortOrder];
+            }
 
             /** @var CodeRepository $repository */
             $repository = $this->getDoctrine()