+++ src/Sulu/Component/Security/Serializer/Subscriber/SecuredEntitySubscriber.php
@@ -1,0 +1,32 @@
+<?php
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Component\Security\Serializer\Subscriber;
+
+use JMS\Serializer\EventDispatcher\EventSubscriberInterface;
+use JMS\Serializer\EventDispatcher\PreSerializeEvent;
+
+class SecuredEntitySubscriber implements EventSubscriberInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public static function getSubscribedEvents()
+    {
+        return [
+            ['event' => 'serializer.pre_serialize', 'method' => 'onPreSerialize'],
+        ];
+    }
+
+    public function onPreSerialize(PreSerializeEvent $event)
+    {
+        $event->getObject();
+    }
+}